Added ratings to maps, and fixed bug with submitting comment replies

+++ src/server/db/migrations/20181105111239_add_ratings_to_maps.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
   return knex.schema
   .createTable('map_ratings', function(table) {
     table.increments();
-    table.integer('rating');
+    table.decimal('rating');
     table.string('steam_id').references('users.steam_id').notNullable().onDelete('cascade');
     table.integer('map_id').references('maps.id').notNullable().onDelete('cascade');
   })
